feat: connet postRecruitingOrder and chatting
### 주요내용
 D/VIDE! 누르면 넘어가기(채팅방에 유저 넣기 + 화면 전환)

+++ app/src/main/java/com/d_vide/D_VIDE/DivideAppState.kt
@@ -60,7 +60,7 @@ class DivideAppState(
         }
     }
 
-    fun navigateToChatting(chattingId: Int, from: NavBackStackEntry) {
+    fun navigateToChatting(chattingId: Long, from: NavBackStackEntry) {
         // In order to discard duplicated navigation events, we check the Lifecycle
         if (from.lifecycleIsResumed()) {
             navController.navigate("${Screen.ChattingDetailScreen.route}/$chattingId")
